Move `FnPtr` impls to the `impls` module
